pyproject.toml
[tool.poetry]
name = "bcgdv-progimage"
version = "0.1.0"
description = ""
authors = ["Ong Tze Wei <ong.tzewei@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
python-decouple = "^3.6"
djangorestframework = "^3.13.1"
Pillow = "^9.1.1"
drf-extra-fields = "^3.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
